do not allow creating multiple new profiles (#216833)
